+++ src/visiomode/models.py
@@ -94,7 +94,6 @@ class Session(Base):
     """Session model class.
 
     Attributes:
-        version: Visiomode version this was generated with.
         animal_id: String representing the animal identifier.
         experiment: A string holding the experiment identifier.
         protocol: An instance of the Protocol class.
@@ -103,7 +102,9 @@ class Session(Base):
         timestamp: A string with the session start date and time (ISO format). Defaults to current date and time.
         notes: String with additional session notes. Defaults to empty string
         device: String hostname of the device running the session. Defaults to the hostname provided by the socket lib.
-        trials: A mutable list of session trials; each trial is an instance of the Trial dataclass.
+        trials: A mutable list of session trials; each trial is an instance of the Trial dataclass. Automatically populated using protocol.trials after class instantiation.
+        animal_meta: A dictionary with animal metadata (see Animal class). Automatically populated using animal_id after class instantiation.
+        version: Visiomode version this was generated with.
     """
 
     animal_id: str
@@ -116,9 +117,11 @@ class Session(Base):
     notes: str = ""
     device: str = socket.gethostname()
     trials: typing.List[Trial] = dataclasses.field(default_factory=list)
+    animal_meta: dict = None
     version: str = __about__.__version__
 
     def __post_init__(self):
+        self.animal_meta = Animal.get_animal(self.animal_id)
         self.trials = self.protocol.trials
 
     def to_dict(self):
@@ -163,6 +166,7 @@ class Animal(Base):
         species: String representing the animal's species. Use the latin name, eg. Mus musculus.
         genotype: String representing the animal's genotype. Defaults to empty string.
         description: String with additional animal notes. Defaults to empty string.
+        rfid: String representing the animal's RFID tag. Defaults to empty string.
     """
 
     animal_id: str
@@ -171,6 +175,7 @@ class Animal(Base):
     species: str
     genotype: str = ""
     description: str = ""
+    rfid: str = ""
 
     def save(self):
         """Append animal to json database file."""
@@ -193,7 +198,20 @@ class Animal(Base):
                 json.dump([self.to_dict()], f)
 
     @classmethod
-    def get_animals(self):
+    def get_animal(cls, animal_id):
+        """Get an animal from the database based on its ID."""
+        path = cfg.db_dir + os.sep + "animals.json"
+
+        if os.path.exists(path):
+            with open(path, "r") as f:
+                animals = json.load(f)
+            for animal in animals:
+                if animal["animal_id"] == animal_id:
+                    return animal
+        return None
+
+    @classmethod
+    def get_animals(cls):
         """Get all animals stored in the database.
 
         Returns:
@@ -208,7 +226,7 @@ class Animal(Base):
         return []
 
     @classmethod
-    def delete_animal(self, animal_id):
+    def delete_animal(cls, animal_id):
         """Delete animal from database."""
         path = cfg.db_dir + os.sep + "animals.json"