+++ src/visiomode/webpanel/api.py
@@ -1,4 +1,5 @@
 """API Module"""
+
 #  This file is part of visiomode.
 #  Copyright (c) 2020 Constantinos Eleftheriou <Constantinos.Eleftheriou@ed.ac.uk>
 #  Distributed under the terms of the MIT Licence.
@@ -155,7 +156,12 @@ class AnimalsAPI(flask.views.MethodView):
         print(request)
         if request_type == "delete":
             animal_id = request.get("id")
-            Animal.delete_animal(animal_id)
+            if animal_id:
+                Animal.delete_animal(animal_id)
+            else:
+                animals = Animal.get_animals()
+                for animal in animals:
+                    Animal.delete_animal(animal["animal_id"])
         elif (request_type == "update") or (request_type == "add"):
             animal = Animal(
                 animal_id=request.get("id"),
@@ -164,6 +170,7 @@ class AnimalsAPI(flask.views.MethodView):
                 species=request.get("species"),
                 genotype=request.get("genotype"),
                 description=request.get("description"),
+                rfid=request.get("rfid"),
             )
             animal.save()
         return json.dumps({"success": True}), 200, {"ContentType": "application/json"}