+++ src/visiomode/webpanel/static/js/session.js
@@ -243,7 +243,7 @@ function isCanvasBlank(canvas) {
 // });
 
 
-/// Dynamic form updates for protocol selection
+/// Dynamic form updates
 
 // load protocol options on select
 protocol_selector = document.getElementById('protocol');
@@ -255,3 +255,62 @@ protocol_selector.onchange = function () {
 }
 
 protocol_selector.onchange();
+
+
+// load animals
+function loadAnimals() {
+    let animal_selector = document.getElementById('animal_id');
+    return $.get("/api/animals").done(function (data) {
+        animal_selector.innerHTML = "";
+        data.animals.reverse(); // reverse order to show latest animals first
+        data.animals.forEach(function (animal) {
+            let option = document.createElement('option');
+            option.value = animal.animal_id;
+            option.text = animal.animal_id;
+            animal_selector.add(option);
+        });
+    });
+}
+
+loadAnimals();
+
+
+// Modals
+
+let addAnimalButton = document.getElementById('add-animal-btn');
+
+function addAnimal() {
+    let animalId = document.getElementById("new-animal-id").value;
+    let animalDob = document.getElementById("new-animal-dob").value;
+    let animalSex = document.getElementById("new-animal-sex").value;
+    let animalSpecies = document.getElementById("new-animal-species").value;
+    let animalGenotype = document.getElementById("new-animal-genotype").value;
+    let animalDescription = document.getElementById("new-animal-description").value;
+    let animalRFID = document.getElementById("new-animal-rfid").value;
+
+    return $.ajax({
+        type: 'POST',
+        url: "/api/animals",
+        data: JSON.stringify({
+            type: "add",
+            data: {
+                id: animalId,
+                dob: animalDob,
+                sex: animalSex,
+                species: animalSpecies,
+                genotype: animalGenotype,
+                description: animalDescription,
+                rfid: animalRFID,
+            },
+        }),
+        dataType: "json",
+        contentType: "application/json",
+        success: function () {
+            $("#addAnimal").modal("hide");
+        }
+    });
+}
+
+addAnimalButton.onclick = function () {
+    addAnimal().done(loadAnimals);
+}