+++ src/visiomode/webpanel/static/js/settings.js
@@ -12,7 +12,10 @@ let clearCacheButton = document.getElementById('clear-cache-btn');
 let deleteAllDataButton = document.getElementById('delete-all-data-btn');
 
 let addAnimalButton = document.getElementById('add-animal-btn');
+let deleteAnimalDataButton = document.getElementById('delete-animal-data-btn');
 
+
+/// Display & storage settings
 
 function loadSettings() {
     $.get("/api/settings", function (data) {
@@ -42,6 +45,26 @@ function updateSettings() {
     });
 }
 
+displaySettingsButton.onclick = function () {
+    currentSettings.width = Number(document.getElementById("display-width").value);
+    currentSettings.height = Number(document.getElementById("display-height").value);
+    currentSettings.fps = Number(document.getElementById("display-fps").value);
+    currentSettings.fullscreen = document.getElementById("display-fullscreen").value === "true";
+
+    updateSettings();
+}
+
+storageSettingsButton.onclick = function () {
+    currentSettings.data_dir = document.getElementById("storage-path").value;
+
+    updateSettings();
+}
+
+loadSettings();
+
+
+/// Animals
+
 function addAnimal() {
     let animalId = document.getElementById("animal-id").value;
     let animalDob = document.getElementById("animal-dob").value;
@@ -49,6 +72,7 @@ function addAnimal() {
     let animalSpecies = document.getElementById("animal-species").value;
     let animalGenotype = document.getElementById("animal-genotype").value;
     let animalDescription = document.getElementById("animal-description").value;
+    let animalRFID = document.getElementById("animal-rfid").value;
 
     $.ajax({
         type: 'POST',
@@ -62,6 +86,7 @@ function addAnimal() {
                 species: animalSpecies,
                 genotype: animalGenotype,
                 description: animalDescription,
+                rfid: animalRFID,
             },
         }),
         dataType: "json",
@@ -72,25 +97,46 @@ function addAnimal() {
     });
 }
 
-displaySettingsButton.onclick = function () {
-    currentSettings.width = Number(document.getElementById("display-width").value);
-    currentSettings.height = Number(document.getElementById("display-height").value);
-    currentSettings.fps = Number(document.getElementById("display-fps").value);
-    currentSettings.fullscreen = document.getElementById("display-fullscreen").value === "true";
-
-    updateSettings();
-}
-
-
-storageSettingsButton.onclick = function () {
-    currentSettings.data_dir = document.getElementById("storage-path").value;
-
-    updateSettings();
-}
-
 addAnimalButton.onclick = function () {
     addAnimal();
 }
 
+function exportAnimals() {
+    // Export animals as CSV
+    $.get("/api/animals", function (data) {
+        animals = data.animals;
+        let replacer = (key, value) => value === null ? '' : value
+        let header = Object.keys(animals[0])
+        let csv = [
+            header.join(','), // header row first
+            ...animals.map(row => header.map(
+                fieldName => JSON.stringify(row[fieldName], replacer)).join(','))
+        ].join('\r\n');
+        let csvExport = new Blob([csv], {type: "text/csv"});
+        let url = window.URL.createObjectURL(csvExport);
+        let a = document.createElement('a');
+        a.href = url;
+        a.download = 'animals.csv';
+        a.click();
+    });
+}
 
-loadSettings();
+function deleteAnimalData() {
+    $.ajax({
+        type: 'POST',
+        url: "/api/animals",
+        data: JSON.stringify({
+            type: "delete",
+            data: {},
+        }),
+        dataType: "json",
+        contentType: "application/json",
+        success: function (data) {
+            $("#deleteAnimalData").modal("hide");
+        }
+    });
+}
+
+deleteAnimalDataButton.onclick = function () {
+    deleteAnimalData();
+}